### 2025-07-01 #8518 @editorialbot
[PRE REVIEW]: BONSAI: A framework for processing and analysing Electronic Health Records (EHR) data using transformer-based models
**Submitting author:** @Montgomeryyyy (<a href="http://orcid.org/0009-0000-3809-0513">Maria Elkjær Montgomery</a>)
**Repository:** https://github.com/FGA-DIKU/EHR
**Branch with paper.md** (empty if default branch):
**Version:** v2.0.0
