+++ corebehrt/configs/create_data.yaml
@@ -17,6 +17,8 @@ features:
   values:
     value_creator_kwargs:
       num_bins: 100
+      add_prefix: true
+      # prefix_regex: r"^([^/]+)/"
   agg_kwargs:
     regex: ^(?:LAB).* # example regex to aggregate all features that start with 'LAB'
     agg_func: "first"
