+++ tests/test_modules/test_values.py
@@ -49,7 +49,9 @@ class TestCreators(unittest.TestCase):
 
         # Create sample data as pandas DataFrames
         self.concepts_pd_normed = self._create_concepts(self.lab_dict_normed)
-        self.concepts_pd_normed_prefix = self._create_concepts(self.lab_dict_normed_prefix)
+        self.concepts_pd_normed_prefix = self._create_concepts(
+            self.lab_dict_normed_prefix
+        )
 
     def test_create_binned_value(self):
         binned_values = ValueCreator.bin_results(self.concepts_pd_normed, num_bins=100)
@@ -73,16 +75,27 @@ class TestCreators(unittest.TestCase):
         self.assertEqual(sorted_concepts, expected_flattened_binned_concepts)
 
     def test_create_binned_value_with_prefix(self):
-        binned_values = ValueCreator.bin_results(self.concepts_pd_normed_prefix, num_bins=100, add_prefix=True, prefix_regex=r"^([^/]+)/")
+        binned_values = ValueCreator.bin_results(
+            self.concepts_pd_normed_prefix,
+            num_bins=100,
+            add_prefix=True,
+            prefix_regex=r"^([^/]+)/",
+        )
         sorted_concepts = list(
             binned_values.sort_values(by=["index", "order"]).sort_index()["code"]
         )
         expected_binned_concepts = [
             [
                 lab,
-                prefix + "VAL_" + str(int(float(value) * 100)) if value != "Kommentar" else None,
+                prefix + "VAL_" + str(int(float(value) * 100))
+                if value != "Kommentar"
+                else None,
             ]
-            for prefix, lab, values in zip(["S/", "S/", "L/", "L/", "L/"], self.lab_dict_normed_prefix.keys(), self.lab_dict_normed_prefix.values())
+            for prefix, lab, values in zip(
+                ["S/", "S/", "L/", "L/", "L/"],
+                self.lab_dict_normed_prefix.keys(),
+                self.lab_dict_normed_prefix.values(),
+            )
             for value in values
         ]
         expected_flattened_binned_concepts = [
