### 2025-07-01 0f87449
remove hard-corded regex
