### 2025-07-02 afd2c57
added option to add prefix to val (#305)
* added option to add prefix to val

* format

* remove from config

* remove hard-corded regex

* regex check

* add non-matching prefix

* name change
