+++ corebehrt/synthetic_data/simulate_synthetic_labs.py
@@ -11,33 +11,49 @@ import argparse
 from pathlib import Path
 
 # Default parameters
-PATIENTS_INFO_PATH = "data/patients_info_20percent.parquet"
-DEFAULT_WRITE_DIR = "tmp/binary_labs"
+PATIENTS_INFO_PATH = "../../data/vals/patient_infos/patient_info_100n.parquet"
+DEFAULT_WRITE_DIR = "../../data/vals/synthetic_data/100n/"
+SAVE_NAME = 'bn_labs_n100_50p_1unq'
 
 # Define lab value ranges and their probabilities for different conditions
 LAB_VALUE_INFO = {
     "S/LAB1": {
-        "positive": {
+        "high": {
+            # "dist": "normal",
+            # "mean": 0.6,
+            # "std": 0.05,
             "dist": "uniform",
-            "range": [1],
-            "n_samples": [1, 3]
+            "range": [1],  # High cases get value 1
+            "n_samples": [1]  # Only one sample per patient
         },
-        "negative": {
-            "mean": 100,
-            "std": 25,
-            "dist": "normal",
-            "n_samples": [1, 3]
+        "low": {
+            # "dist": "normal",
+            # "mean": 0.4,
+            # "std": 0.05,
+            "dist": "uniform",
+            "range": [0],  # Low cases get value 0
+            "n_samples": [1]  # Only one sample per patient
         },
     },
 }
 
 CONCEPT_RELATIONSHIPS = {
-    "S/DIAG1": {
-        "base_probability": 1.0,
+    "S/LAB1": {
+        "base_probability": 1.0,  # 100% of patients get labs
+        "condition_probabilities": {
+            "high": 0.5,  # 50% chance of being high
+            "low": 0.5,   # 50% chance of being low
+        },
+        "add_base_concept": ["high", "low"],  # Add lab for all conditions
         "related_concepts": {
-            "S/LAB1": {
-                "prob": 0.5,
-                "conditions": ["positive"],
+            "S/DIAG1": {
+                "prob": 1,  # 100% chance of getting diagnosis if lab is high
+                "conditions": ["high"],  # Only high lab patients get diagnosis
+                "time_relationship": {
+                    "type": "after",  # Diagnosis comes after lab
+                    "min_days": 1,
+                    "max_days": 180
+                }
             }
         }
     },
@@ -79,18 +95,22 @@ def get_related_concepts(base_concept: str) -> Tuple[List[str], List[float]]:
     
     if base_concept not in CONCEPT_RELATIONSHIPS:
         return related_concepts, values
-        
+    
+    # Determine the condition based on probabilities
+    condition_probs = CONCEPT_RELATIONSHIPS[base_concept]["condition_probabilities"]
+    conditions = list(condition_probs.keys())
+    probs = list(condition_probs.values())
+    condition = np.random.choice(conditions, p=probs)
+    
     for lab_name, info in CONCEPT_RELATIONSHIPS[base_concept]["related_concepts"].items():
-        if np.random.random() < info["prob"]:
-            # Generate condition for patient
-            condition = np.random.choice(info["conditions"])
-            n_samples = np.random.randint(LAB_VALUE_INFO[lab_name][condition]["n_samples"][0], LAB_VALUE_INFO[lab_name][condition]["n_samples"][1])
-            
-            for _ in range(n_samples):
-                value = generate_lab_value(lab_name, condition)
-                if value is not None:
-                    related_concepts.append(lab_name)
-                    values.append(value)
+        # Check if we should generate a lab value
+        should_generate_lab = np.random.random() < info["prob"]
+        if should_generate_lab:
+            # Generate value based on the selected condition
+            value = generate_lab_value(lab_name, condition)
[…]
