+++ corebehrt/synthetic_data/modify_synthetic_labs.py
@@ -1,0 +1,293 @@
+"""
+Modify lab values in an existing synthetic data CSV file according to new distributions.
+"""
+
+import pandas as pd
+import numpy as np
+import argparse
+from pathlib import Path
+from typing import Optional
+import matplotlib.pyplot as plt
+import seaborn as sns
+import os
+from theoretical_separation import cohens_d, sweep_threshold_auc, manual_mann_whitney_u, scipy_mann_whitney_u
+
+# Default parameters
+N = 100000
+LOW_MEAN = 0.48
+HIGH_MEAN = 0.52
+STD = 0.10
+DEFAULT_INPUT_FILE = f"../../data/vals/synthetic_data/{N}n/bn_labs_n{N}_50p_1unq.csv"
+DEFAULT_WRITE_DIR = f"../../data/vals/synthetic_data/{N}n/" 
+DEFAULT_PLOT_DIR = f"../../data/vals/synthetic_data_plots/{N}n/"
+SAVE_NAME = f'gaus_mean_{LOW_MEAN}_{HIGH_MEAN}_std_{STD}_n{N}_50p_1unq'
+POSITIVE_DIAGS = ['S/DIAG1']
+
+# Define new lab value distributions
+LAB_VALUE_INFO = {
+    "S/LAB1": {
+        "high": {
+            "dist": "normal",
+            "mean": HIGH_MEAN,
+            "std": STD,
+        },
+        "low": {
+            "dist": "normal",
+            "mean": LOW_MEAN,
+            "std": STD,
+        },
+    },
+}
+
+def generate_lab_value(lab_name: str, condition: str) -> Optional[float]:
+    """
+    Generate a lab value based on the lab name and condition.
+    
+    Args:
+        lab_name: Name of the lab test
+        condition: The condition affecting the lab values
+        
+    Returns:
+        Optional[float]: Generated lab value or None if invalid input
+    """
+    if lab_name not in LAB_VALUE_INFO or condition not in LAB_VALUE_INFO[lab_name]:
+        return None
+        
+    range_info = LAB_VALUE_INFO[lab_name][condition]
+    if range_info["dist"] == "uniform":
+        return np.random.choice(range_info["range"])
+    elif range_info["dist"] == "normal":
+        return np.random.normal(range_info["mean"], range_info["std"])
+    return None
+
+def modify_lab_values(data: pd.DataFrame) -> pd.DataFrame:
+    """
+    Modify only the numeric values of existing lab entries while keeping everything else the same.
+    
+    Args:
+        data: DataFrame containing the synthetic data
+        
+    Returns:
+        pd.DataFrame: Modified DataFrame with updated lab values
+    """
+    # Create a copy of the data
+    modified_data = data.copy()
+    
+    # Get the mask for lab entries
+    lab_mask = modified_data['code'] == 'S/LAB1'
+    
+    # For each lab entry, generate a new value based on the patient's condition
+    for idx in modified_data[lab_mask].index:
+        patient_id = modified_data.at[idx, 'subject_id']
+        is_positive = modified_data.at[idx, 'is_positive']
+        condition = 'high' if is_positive else 'low'
+        
+        # Generate new value
+        new_value = generate_lab_value('S/LAB1', condition)
+        if new_value is not None:
+            modified_data.at[idx, 'numeric_value'] = new_value
+    
+    return modified_data
+
+def get_positive_patients(data: pd.DataFrame, positive_diags: list) -> pd.DataFrame:
+    """
+    Get positive patients from the data and add is_positive column.
+    
+    Args:
+        data: DataFrame containing the synthetic data
+        positive_diags: List of diagnosis codes that indicate positive cases
+        
[…]
