+++ corebehrt/azure/pipelines/__init__.py
@@ -1,4 +1,6 @@
 from corebehrt.azure.pipelines.E2E import E2E
+from corebehrt.azure.pipelines.E2E_full import E2E_full
 from corebehrt.azure.pipelines.FINETUNE import FINETUNE
 
-PIPELINE_REGISTRY = [E2E, FINETUNE]
+
+PIPELINE_REGISTRY = [E2E, E2E_full, FINETUNE]
