### 2025-07-04 96d522e
minimum_index_date
