### 2025-07-04 0d8a2b9
fix outcome path
