### 2025-07-05 f6585f2
metrics logging xgb
