+++ corebehrt/azure/pipelines/__init__.py
@@ -1,6 +1,7 @@
 from corebehrt.azure.pipelines.E2E import E2E
 from corebehrt.azure.pipelines.E2E_full import E2E_full
 from corebehrt.azure.pipelines.FINETUNE import FINETUNE
+from corebehrt.azure.pipelines.E2E_XGB import E2E_XGB
 
 
-PIPELINE_REGISTRY = [E2E, E2E_full, FINETUNE]
+PIPELINE_REGISTRY = [E2E, E2E_full, FINETUNE, E2E_XGB]
