### 2025-07-09 20cc800
remove final separator token if present
