### 2025-07-09 23c0f54
remove final separator token if present (#306)
* remove final separator token if present

* format
