### 2025-07-28 dac9b7b
more synthetic data creation
