+++ corebehrt/synthetic_data/histogram_equalisation.py
@@ -1,0 +1,322 @@
+"""
+Apply histogram equalisation to lab values in synthetic data CSV files.
+"""
+
+import pandas as pd
+import numpy as np
+import argparse
+from pathlib import Path
+import matplotlib.pyplot as plt
+import os
+from scipy import stats
+from theoretical_separation import cohens_d, sweep_threshold_auc, scipy_mann_whitney_u
+
+# Default parameters
+N = 10000
+LOW_MEAN = 0.45
+HIGH_MEAN = 0.55
+STD = 0.15
+N_BINS = 256
+DEFAULT_INPUT_FILE = f"../../../data/vals/synthetic_data/{N}n/gaus_mean_{int(LOW_MEAN*100)}_{int(HIGH_MEAN*100)}_std_{int(STD*100)}_n{N}_50p_1unq.csv"
+DEFAULT_WRITE_DIR = f"../../../data/vals/synthetic_data/{N}n/"
+DEFAULT_PLOT_DIR = f"../../../data/vals/synthetic_data_plots/{N}n/"
+SAVE_NAME = f"gaus_mean_{int(LOW_MEAN*100)}_{int(HIGH_MEAN*100)}_std_{int(STD*100)}_n{N}_50p_1unq_histeq_n{N_BINS}"
+
+def histogram_equalisation(values: np.ndarray, n_bins: int = 256) -> np.ndarray:
+    """
+    Perform histogram equalisation on a 1D array of values.
+    
+    Args:
+        values: Input array of values to equalise
+        n_bins: Number of bins for histogram calculation (default: 256)
+        
+    Returns:
+        np.ndarray: Equalized values
+    """
+    # Calculate histogram
+    hist, bin_edges = np.histogram(values, bins=n_bins, density=True)
+    
+    # Calculate cumulative distribution function
+    cdf = hist.cumsum()
+    cdf = cdf / cdf[-1]  # Normalize to [0, 1]
+    
+    # Create interpolation function
+    bin_centers = (bin_edges[:-1] + bin_edges[1:]) / 2
+    
+    # Interpolate to get equalised values
+    equalised_values = np.interp(values, bin_centers, cdf)
+    
+    return equalised_values
+
+def apply_histogram_equalisation(data: pd.DataFrame, lab_code: str = 'S/LAB1', n_bins: int = 256) -> pd.DataFrame:
+    """
+    Apply histogram equalisation to lab values in the dataset.
+    
+    Args:
+        data: DataFrame containing the synthetic data
+        lab_code: The lab code to equalise
+        n_bins: Number of bins for histogram calculation
+        
+    Returns:
+        pd.DataFrame: DataFrame with equalised lab values
+    """
+    # Create a copy of the data
+    equalised_data = data.copy()
+    
+    # Get the mask for lab entries
+    lab_mask = equalised_data['code'] == lab_code
+    
+    if not lab_mask.any():
+        print(f"Warning: No entries found for lab code '{lab_code}'")
+        return equalised_data
+    
+    # Get the lab values
+    lab_values = equalised_data.loc[lab_mask, 'numeric_value'].values
+    
+    # Apply histogram equalisation
+    equalised_values = histogram_equalisation(lab_values, n_bins)
+    
+    # Update the data
+    equalised_data.loc[lab_mask, 'numeric_value'] = equalised_values
+    
+    return equalised_data
+
+def print_statistics(data: pd.DataFrame, title: str = "Lab value statistics") -> None:
+    """
+    Print statistics about the lab values.
+    
+    Args:
+        data: DataFrame containing the data
+        title: Title for the statistics section
+    """
+    # Get lab values for positive and negative patients
+    lab_mask = data['code'] == 'S/LAB1'
+    positive_mask = data['is_positive']
+    
[…]
