+++ corebehrt/synthetic_data/add_noise.py
@@ -1,0 +1,431 @@
+"""
+Add noise to synthetic data by switching labels and optionally removing data.
+"""
+
+import pandas as pd
+import numpy as np
+import argparse
+from pathlib import Path
+import matplotlib.pyplot as plt
+import os
+from theoretical_separation import cohens_d, sweep_threshold_auc, manual_mann_whitney_u, scipy_mann_whitney_u
+
+# Default parameters
+N = 10000
+LOW_MEAN = 0.45
+HIGH_MEAN = 0.55
+STD = 0.10
+DEFAULT_INPUT_FILE = f"../../../data/vals/synthetic_data/{N}n/gaus_mean_{int(LOW_MEAN*100)}_{int(HIGH_MEAN*100)}_std_{int(STD*100)}_n{N}_50p_1unq.csv"
+DEFAULT_WRITE_DIR = f"../../../data/vals/synthetic_data/{N}n/"
+DEFAULT_PLOT_DIR = f"../../../data/vals/synthetic_data_plots/{N}n/"
+
+# Default noise parameters
+DEFAULT_SWITCH_PERCENTAGE = 0.0  # 5% label switching
+DEFAULT_REMOVE_PERCENTAGE = 0.05 #0   # 0% data removal
+DEFAULT_SEED = 42                 # Fixed seed for reproducibility
+
+def add_label_noise(data: pd.DataFrame, switch_percentage: float = 0.05) -> pd.DataFrame:
+    """
+    Add noise by switching labels for a percentage of positive and negative cases.
+    This is done by swapping diagnostic codes (add S/DIAG1 to negatives, remove from positives).
+    
+    Args:
+        data: DataFrame containing the synthetic data
+        switch_percentage: Percentage of positive and negative cases to switch (0.0 to 1.0)
+        
+    Returns:
+        pd.DataFrame: DataFrame with noisy labels
+    """
+    noisy_data = data.copy()
+    
+    # Get positive and negative patients
+    positive_patients = set(noisy_data[noisy_data['is_positive']]['subject_id'].unique())
+    negative_patients = set(noisy_data[~noisy_data['is_positive']]['subject_id'].unique())
+    
+    # Calculate number of patients to switch
+    n_positive_to_switch = int(len(positive_patients) * switch_percentage)
+    n_negative_to_switch = int(len(negative_patients) * switch_percentage)
+    
+    # Randomly select patients to switch
+    positive_to_switch = set(np.random.choice(list(positive_patients), n_positive_to_switch, replace=False))
+    negative_to_switch = set(np.random.choice(list(negative_patients), n_negative_to_switch, replace=False))
+    
+    # Swap diagnostic codes (add S/DIAG1 to negatives, remove from positives)
+    print(f"Swapping diagnostic codes for {len(positive_to_switch)} positive and {len(negative_to_switch)} negative patients...")
+    
+    # Remove S/DIAG1 from positive patients being switched
+    for patient_id in positive_to_switch:
+        # Remove diagnostic code
+        diag_mask = (noisy_data['subject_id'] == patient_id) & (noisy_data['code'] == 'S/DIAG1')
+        noisy_data = noisy_data[~diag_mask]
+        # Update label
+        noisy_data.loc[noisy_data['subject_id'] == patient_id, 'is_positive'] = False
+    
+    # Add S/DIAG1 to negative patients being switched
+    for patient_id in negative_to_switch:
+        # Add diagnostic code (assuming we have the same structure as existing DIAG1 records)
+        existing_diag = noisy_data[noisy_data['code'] == 'S/DIAG1'].iloc[0] if len(noisy_data[noisy_data['code'] == 'S/DIAG1']) > 0 else None
+        if existing_diag is not None:
+            new_diag_row = existing_diag.copy()
[…]
