### 2025-07-29 26fa4cf
minimum index dates
