+++ corebehrt/modules/cohort_handling/index_dates.py
@@ -109,6 +109,7 @@ class IndexDateHandler:
         n_hours_from_exposure: Optional[int] = None,
         exposures: Optional[pd.DataFrame] = None,
         minimum_index_dates: Optional[pd.DataFrame] = None,
+        n_hours_from_minimum_index_date: Optional[int] = None,
     ) -> pd.Series:
         """Determine index dates based on mode.
         Args:
@@ -130,6 +131,10 @@ class IndexDateHandler:
             exposed_timestamps = cls.get_index_timestamps_for_exposed(
                 pids, n_hours, exposures
             )
+            if n_hours_from_minimum_index_date is not None:
+                minimum_index_dates = cls.get_index_timestamps_for_exposed(
+                    pids, n_hours_from_minimum_index_date, minimum_index_dates
+                )
             result = cls.draw_index_dates_for_unexposed(pids, exposed_timestamps, minimum_index_dates)
         else:
             raise ValueError(f"Unsupported index date mode: {index_date_mode}")
