### 2025-07-29 17bb7b3
added hours from max
