+++ corebehrt/modules/cohort_handling/index_dates.py
@@ -28,52 +28,59 @@ class IndexDateHandler:
         """Get index timestamps for exposed patients."""
         hours_delta = pd.Timedelta(hours=n_hours_from_exposure)
         exposures = filter_table_by_pids(exposures, pids)
-        # Set PID as index and get timestamp series
         result = exposures.set_index(PID_COL)[TIMESTAMP_COL] + hours_delta
-        result.index.name = PID_COL  # Ensure index name is set to PID_COL
+        result.index.name = PID_COL
         return result
+
+    @staticmethod
+    def _ensure_series_format(data: pd.Series | pd.DataFrame) -> pd.Series:
+        """Ensure data is in Series format with PID as index."""
+        if isinstance(data, pd.DataFrame):
+            return data.set_index(PID_COL)[TIMESTAMP_COL]
+        return data
 
     @staticmethod
     def draw_index_dates_for_unexposed_w_secondary_censoring(
         data_pids: List[str],
         censoring_timestamps: pd.Series,
-        secondary_censoring_timestamps: Optional[pd.DataFrame] = None,
+        secondary_censoring_timestamps: pd.Series,
     ) -> pd.Series:
-        """Draw censor dates for patients not in censor_timestamps."""
+        """
+        Draw censor dates for patients not in censoring_timestamps.
+        Uses secondary censoring timestamps for missing patients when available.
+        """
         np.random.seed(42)
-        missing_pids = set(data_pids) - set(censoring_timestamps.index)
         
-        # Initialize result with existing censoring timestamps
+        # Ensure proper Series format
+        censoring_timestamps = IndexDateHandler._ensure_series_format(censoring_timestamps)
+        secondary_censoring_timestamps = IndexDateHandler._ensure_series_format(secondary_censoring_timestamps)
+        
+        missing_pids = set(data_pids) - set(censoring_timestamps.index)
         result = censoring_timestamps.copy()
         
-        # For patients with secondary censoring timestamps, use those instead
-        if secondary_censoring_timestamps is not None:
-            # Get patients that have secondary censoring timestamps
-            secondary_pids = set(secondary_censoring_timestamps.index) & set(data_pids)
-            
-            # Update result with secondary censoring timestamps for those patients
-            for pid in secondary_pids:
-                if pid in result.index:
-                    # Replace existing censoring timestamp with secondary one
-                    result[pid] = secondary_censoring_timestamps.loc[pid, TIMESTAMP_COL]
-                else:
-                    # Add new entry with secondary censoring timestamp
-                    result[pid] = secondary_censoring_timestamps.loc[pid, TIMESTAMP_COL]
-            
-            # Remove these patients from missing_pids since they now have censoring dates
-            missing_pids = missing_pids - secondary_pids
-        
-        # For remaining missing patients, draw from normal censoring timestamps
         if missing_pids:
-            random_abspos = np.random.choice(
-                censoring_timestamps.values, size=len(missing_pids)
-            )
-            new_entries = pd.Series(
-                random_abspos, index=pd.Index(list(missing_pids), name=PID_COL)
-            )
-            result = pd.concat([result, new_entries])
+            # Get patients that have secondary censoring timestamps
+            secondary_pids = set(secondary_censoring_timestamps.index) & missing_pids
+            
+            # Use secondary timestamps directly for patients that have them
+            for pid in secondary_pids:
+                result[pid] = secondary_censoring_timestamps.loc[pid]
+            
+            # For remaining missing patients, draw randomly from secondary censoring dates
+            remaining_missing = missing_pids - secondary_pids
+            if remaining_missing and len(secondary_censoring_timestamps) > 0:
+                random_secondary_dates = np.random.choice(
+                    secondary_censoring_timestamps.values, 
[…]
