+++ corebehrt/synthetic_data/add_noise.py
@@ -11,7 +11,7 @@ import os
 from theoretical_separation import cohens_d, sweep_threshold_auc, manual_mann_whitney_u, scipy_mann_whitney_u
 
 # Default parameters
-N = 10000
+N = 100000
 LOW_MEAN = 0.45
 HIGH_MEAN = 0.55
 STD = 0.10
@@ -20,14 +20,14 @@ DEFAULT_WRITE_DIR = f"../../../data/vals/synthetic_data/{N}n/"
 DEFAULT_PLOT_DIR = f"../../../data/vals/synthetic_data_plots/{N}n/"
 
 # Default noise parameters
-DEFAULT_SWITCH_PERCENTAGE = 0.0  # 5% label switching
-DEFAULT_REMOVE_PERCENTAGE = 0.05 #0   # 0% data removal
+DEFAULT_SWITCH_PERCENTAGE = 0.0  # % label switching
+DEFAULT_REMOVE_PERCENTAGE = 0.25 #0  % data removal
 DEFAULT_SEED = 42                 # Fixed seed for reproducibility
 
 def add_label_noise(data: pd.DataFrame, switch_percentage: float = 0.05) -> pd.DataFrame:
     """
     Add noise by switching labels for a percentage of positive and negative cases.
-    This is done by swapping diagnostic codes (add S/DIAG1 to negatives, remove from positives).
+    This is done by swapping lab values between positive and negative patients.
     
     Args:
         data: DataFrame containing the synthetic data
@@ -50,31 +50,37 @@ def add_label_noise(data: pd.DataFrame, switch_percentage: float = 0.05) -> pd.D
     positive_to_switch = set(np.random.choice(list(positive_patients), n_positive_to_switch, replace=False))
     negative_to_switch = set(np.random.choice(list(negative_patients), n_negative_to_switch, replace=False))
     
-    # Swap diagnostic codes (add S/DIAG1 to negatives, remove from positives)
-    print(f"Swapping diagnostic codes for {len(positive_to_switch)} positive and {len(negative_to_switch)} negative patients...")
+    print(f"Swapping lab values for {len(positive_to_switch)} positive and {len(negative_to_switch)} negative patients...")
     
-    # Remove S/DIAG1 from positive patients being switched
-    for patient_id in positive_to_switch:
-        # Remove diagnostic code
-        diag_mask = (noisy_data['subject_id'] == patient_id) & (noisy_data['code'] == 'S/DIAG1')
-        noisy_data = noisy_data[~diag_mask]
-        # Update label
-        noisy_data.loc[noisy_data['subject_id'] == patient_id, 'is_positive'] = False
+    # Create pairs of patients to swap lab values
+    positive_switch_list = list(positive_to_switch)
+    negative_switch_list = list(negative_to_switch)
     
-    # Add S/DIAG1 to negative patients being switched
-    for patient_id in negative_to_switch:
-        # Add diagnostic code (assuming we have the same structure as existing DIAG1 records)
-        existing_diag = noisy_data[noisy_data['code'] == 'S/DIAG1'].iloc[0] if len(noisy_data[noisy_data['code'] == 'S/DIAG1']) > 0 else None
-        if existing_diag is not None:
-            new_diag_row = existing_diag.copy()
-            new_diag_row['subject_id'] = patient_id
-            new_diag_row['is_positive'] = True
-            noisy_data = pd.concat([noisy_data, pd.DataFrame([new_diag_row])], ignore_index=True)
-        # Update label
-        noisy_data.loc[noisy_data['subject_id'] == patient_id, 'is_positive'] = True
+    # Ensure we have equal numbers to swap (take the minimum)
+    n_pairs = min(len(positive_switch_list), len(negative_switch_list))
     
[…]
