+++ corebehrt/synthetic_data/add_noise.py
@@ -8,169 +8,212 @@ import argparse
 from pathlib import Path
 import matplotlib.pyplot as plt
 import os
-from theoretical_separation import cohens_d, sweep_threshold_auc, manual_mann_whitney_u, scipy_mann_whitney_u
+from theoretical_separation import (
+    cohens_d,
+    sweep_threshold_auc,
+    manual_mann_whitney_u,
+    scipy_mann_whitney_u,
+)
 
 # Default parameters
 N = 100000
 LOW_MEAN = 0.45
 HIGH_MEAN = 0.55
 STD = 0.10
-DEFAULT_INPUT_FILE = f"../../../data/vals/synthetic_data/{N}n/gaus_mean_{int(LOW_MEAN*100)}_{int(HIGH_MEAN*100)}_std_{int(STD*100)}_n{N}_50p_1unq.csv"
+DEFAULT_INPUT_FILE = f"../../../data/vals/synthetic_data/{N}n/gaus_mean_{int(LOW_MEAN * 100)}_{int(HIGH_MEAN * 100)}_std_{int(STD * 100)}_n{N}_50p_1unq.csv"
 DEFAULT_WRITE_DIR = f"../../../data/vals/synthetic_data/{N}n/"
 DEFAULT_PLOT_DIR = f"../../../data/vals/synthetic_data_plots/{N}n/"
 
 # Default noise parameters
 DEFAULT_SWITCH_PERCENTAGE = 0.0  # % label switching
-DEFAULT_REMOVE_PERCENTAGE = 0.25 #0  % data removal
-DEFAULT_SEED = 42                 # Fixed seed for reproducibility
+DEFAULT_REMOVE_PERCENTAGE = 0.25  # 0  % data removal
+DEFAULT_SEED = 42  # Fixed seed for reproducibility
 
-def add_label_noise(data: pd.DataFrame, switch_percentage: float = 0.05) -> pd.DataFrame:
+
+def add_label_noise(
+    data: pd.DataFrame, switch_percentage: float = 0.05
+) -> pd.DataFrame:
     """
     Add noise by switching labels for a percentage of positive and negative cases.
     This is done by swapping lab values between positive and negative patients.
-    
+
     Args:
         data: DataFrame containing the synthetic data
         switch_percentage: Percentage of positive and negative cases to switch (0.0 to 1.0)
-        
+
     Returns:
         pd.DataFrame: DataFrame with noisy labels
     """
     noisy_data = data.copy()
-    
+
     # Get positive and negative patients
-    positive_patients = set(noisy_data[noisy_data['is_positive']]['subject_id'].unique())
-    negative_patients = set(noisy_data[~noisy_data['is_positive']]['subject_id'].unique())
-    
+    positive_patients = set(
+        noisy_data[noisy_data["is_positive"]]["subject_id"].unique()
+    )
+    negative_patients = set(
+        noisy_data[~noisy_data["is_positive"]]["subject_id"].unique()
+    )
+
     # Calculate number of patients to switch
     n_positive_to_switch = int(len(positive_patients) * switch_percentage)
     n_negative_to_switch = int(len(negative_patients) * switch_percentage)
-    
+
     # Randomly select patients to switch
-    positive_to_switch = set(np.random.choice(list(positive_patients), n_positive_to_switch, replace=False))
-    negative_to_switch = set(np.random.choice(list(negative_patients), n_negative_to_switch, replace=False))
-    
-    print(f"Swapping lab values for {len(positive_to_switch)} positive and {len(negative_to_switch)} negative patients...")
-    
+    positive_to_switch = set(
+        np.random.choice(list(positive_patients), n_positive_to_switch, replace=False)
+    )
+    negative_to_switch = set(
[…]
