+++ corebehrt/main/helper/select_cohort.py
@@ -53,7 +53,16 @@ def select_cohort(
     """
 
     logger.info("Loading data")
-    patients_info, outcomes, exposures, initial_pids, exclude_pids, minimum_index_dates, maximum_index_dates, secondary_censoring_timestamps = load_data(path_cfg)
+    (
+        patients_info,
+        outcomes,
+        exposures,
+        initial_pids,
+        exclude_pids,
+        minimum_index_dates,
+        maximum_index_dates,
+        secondary_censoring_timestamps,
+    ) = load_data(path_cfg)
 
     # Remove duplicate patient records (keep first occurrence)
     patients_info = patients_info.drop_duplicates(subset=PID_COL, keep="first")
@@ -94,10 +103,16 @@ def select_cohort(
         exposures=exposures,
         minimum_index_dates=minimum_index_dates,
         maximum_index_dates=maximum_index_dates,
-        n_hours_from_minimum_index_date=index_date_cfg[mode].get("n_hours_from_minimum_index_date", None),
-        n_hours_from_maximum_index_date=index_date_cfg[mode].get("n_hours_from_maximum_index_date", None),
+        n_hours_from_minimum_index_date=index_date_cfg[mode].get(
+            "n_hours_from_minimum_index_date", None
+        ),
+        n_hours_from_maximum_index_date=index_date_cfg[mode].get(
+            "n_hours_from_maximum_index_date", None
+        ),
         secondary_censoring_timestamps=secondary_censoring_timestamps,
-        n_hours_from_secondary_censoring_timestamps=index_date_cfg[mode].get("n_hours_from_secondary_censoring_timestamps", None),
+        n_hours_from_secondary_censoring_timestamps=index_date_cfg[mode].get(
+            "n_hours_from_secondary_censoring_timestamps", None
+        ),
     )
 
     # This split is done after index date calculation but before any filtering based on index dates
@@ -150,7 +165,15 @@ def log_patient_num(logger, patients_info):
 
 def load_data(
     path_cfg,
-) -> Tuple[pd.DataFrame, pd.DataFrame, pd.DataFrame, List[str], List[str], Optional[pd.DataFrame], Optional[pd.DataFrame]]:
+) -> Tuple[
+    pd.DataFrame,
+    pd.DataFrame,
+    pd.DataFrame,
+    List[str],
+    List[str],
+    Optional[pd.DataFrame],
+    Optional[pd.DataFrame],
+]:
     """Load patient, outcomes, and exposures data."""
     patients_info = ConceptLoader.read_file(path_cfg.patients_info)
     outcomes = ConceptLoader.read_file(path_cfg.outcome)
@@ -173,7 +196,9 @@ def load_data(
         maximum_index_dates = None
 
     if path_cfg.get("secondary_censoring_timestamps", False):
-        secondary_censoring_timestamps = ConceptLoader.read_file(path_cfg.secondary_censoring_timestamps)
+        secondary_censoring_timestamps = ConceptLoader.read_file(
+            path_cfg.secondary_censoring_timestamps
+        )
     else:
         secondary_censoring_timestamps = None
 
@@ -185,4 +210,13 @@ def load_data(
         torch.load(path_cfg.exclude_pids) if path_cfg.get("exclude_pids", False) else []
     )
 
-    return patients_info, outcomes, exposures, initial_pids, exclude_pids, minimum_index_dates, maximum_index_dates, secondary_censoring_timestamps
+    return (
+        patients_info,
+        outcomes,
+        exposures,
+        initial_pids,
+        exclude_pids,
+        minimum_index_dates,
+        maximum_index_dates,
+        secondary_censoring_timestamps,
+    )
