### 2025-07-30 84e8712
addition of positive/negative diags
