### 2025-08-05 42b167b
removed debug
