### 2025-08-26 #8518 @espottesmith
Very sorry for the delay. I was moving across the Atlantic.

I can take this on, though I'm not particularly knowledgeable on EHR.
