### 2025-08-26 #8869 @editorialbot
[REVIEW]: BONSAI: A framework for processing and analysing Electronic Health Records (EHR) data using transformer-based models
**Submitting author:** @Montgomeryyyy (<a href="http://orcid.org/0009-0000-3809-0513">Maria Elkjær Montgomery</a>)
**Repository:** https://github.com/FGA-DIKU/EHR
**Branch with paper.md** (empty if default branch):
**Version:** 2.0.1
