+++ paper/paper.bib
@@ -86,3 +86,25 @@
   year={2021},
   publisher={Nature Publishing Group UK London}
 }
+
+@incollection{pytorch_2019,
+  title     = {PyTorch: An Imperative Style, High-Performance Deep Learning Library},
+  author    = {Paszke, Adam and Gross, Sam and Massa, Francisco and Lerer, Adam and Bradbury, James and Chanan, Gregory and Killeen, Trevor and Lin, Zeming and Gimelshein, Natalia and Antiga, Luca and Desmaison, Alban and Kopf, Andreas and Yang, Edward and DeVito, Zachary and Raison, Martin and Tejani, Alykhan and Chilamkurthy, Sasank and Steiner, Benoit and Fang, Lu and Bai, Junjie and Chintala, Soumith},
+  booktitle = {Advances in Neural Information Processing Systems 32},
+  pages     = {8024--8035},
+  year      = {2019},
+  publisher = {Curran Associates, Inc.},
+  url       = {http://papers.neurips.cc/paper/9015-pytorch-an-imperative-style-high-performance-deep-learning-library.pdf}
+}
+
+@misc{huggingface_transformers,
+  title = {Transformers},
+  howpublished = {\url{https://huggingface.co/docs/transformers/en/index}},
+  note = {Accessed: 2025-09-25}
+}
+
+@misc{mlflow,
+  title = {mlflow.azureml},
+  howpublished = {\url{https://mlflow.org/docs/1.22.0/python_api/mlflow.azureml.html}},
+  note = {Accessed: 2025-09-25}
+}
