+++ paper/paper.md
@@ -42,16 +42,18 @@ bibliography: paper.bib
 ---
 
 # Summary 
-BONSAI is an end-to-end Python framework for processing and analysing Electronic Health Records (EHR) data. It extends the model described in the CORE-BEHRT paper [@odgaard2024core] and is designed to streamline data preparation, model pre-training, and fine-tuning for patient-level prediction tasks. The framework is built for efficient large-scale processing of EHR data, making it suitable for clinical applications involving substantial volumes of patient record. 
+BONSAI is an end-to-end Python framework for processing and analysing Electronic Health Records (EHR) data. It extends the model described in the CORE-BEHRT paper [@odgaard2024core] and is designed to streamline data preparation, model pre-training, and fine-tuning for patient-level prediction tasks. The framework is built for efficient large-scale processing of EHR data, making it suitable for clinical applications involving substantial volumes of patient record. This enables models for patient outcome prediction and supports scalable clinical research.
 
 The framework accepts EHR data in the MEDS format [@kolo2024meds], performs comprehensive preprocessing, and prepares the data for BERT-based modeling, following the structure of the BEHRT model [@li2020behrt]. This includes converting raw EHR data into tokenised inputs by mapping vocabulary to numerical tokens and aligning patient histories into temporally ordered sequences of medical concepts. Each concept is paired with temporal features such as visit timestamps (positions), patient age, and visit-level segment encodings. Additional static features, including date of birth (DOB), gender, and optionally date of death (DOD), are prepended to each sequence. Separator and classification tokens can be optionally included. Numeric values can be binned and added as categorical tokens. Aggregation of similar concepts is supported via regex-based grouping, and the removal of specific concepts is also included via regex.
 
-Modeling is performed using a ModernBERT backbone [@warner2024smarter] from the Hugging Face library. Model configurations are specified via a YAML config file, with defaults provided. Pre-training uses a masked language modeling (MLM) objective with Cross-Entropy loss. Fine-tuning is performed as a binary classification task using outcome-specific labels derived from the input sequences. Cohort definition and censoring can be done in two ways: either by including all data with post-hoc censoring based on outcome dates and a user-defined window, or via a simulated prospective approach with a fixed cutoff date. The fine-tuning head consists of a BiGRU layer that encodes patient sequences into a single embedding, which is passed to a linear classification layer trained using binary cross-entropy loss and the AdamW optimizer. Optional features include the use of a learning rate scheduler and the ability to freeze pre-trained layers during fine-tuning. Finally, the pipeline also includes an evaluation script that outputs model predictions and, optionally, intermediate model embeddings.
[…]
