### 2025-09-25 0f090bc
white pipeline fig
