+++ paper/paper.md
@@ -62,7 +62,7 @@ Figure \autoref{fig:pipeline} depicts the overall pipeline of BONSAI.
 \autoref{fig:pipeline} depicts the overall pipeline of BONSAI. 
 >>>>>>> 6ea7e61 (name change)
 
-![The BONSAI pipeline.\label{fig:pipeline}](../docs/COREBEHRT_overview_dark.jpg){width=100%}
+![The BONSAI pipeline.\label{fig:pipeline}](../docs/BONSAI_pipeline.jpg){width=100%}
 
 The figures below depict the censoring scheme for the data preprocessing, where \autoref{fig:censoring} A shows the censoring scheme for the post-hoc setup, and \autoref{fig:censoring} B shows the censoring scheme in a prospective setup.
 
