+++ paper/paper.md
@@ -42,7 +42,7 @@ bibliography: paper.bib
 ---
 
 # Summary 
-BONSAI is an end-to-end Python framework for processing and analysing Electronic Health Records (EHR) data. It extends the model described in the CORE-BEHRT paper [@odgaard2024core] and is designed to streamline data preparation, model pre-training, and fine-tuning for patient-level prediction tasks. The framework is built for efficient large-scale processing of EHR data, making it suitable for clinical applications involving substantial volumes of patient record. This enables models for patient outcome prediction and supports scalable clinical research.
+BONSAI is an end-to-end Python framework for processing and analysing Electronic Health Records (EHR) data. It extends the model described in the CORE-BEHRT paper [@odgaard2024core] and is designed to streamline data preparation, model pre-training, and fine-tuning for patient-level prediction tasks. The framework is built for efficient large-scale processing of EHR data, making it suitable for clinical applications involving substantial volumes of patient records. This enables models for patient outcome prediction and supports scalable clinical research.
 
 The framework accepts EHR data in the MEDS format [@kolo2024meds], performs comprehensive preprocessing, and prepares the data for BERT-based modeling, following the structure of the BEHRT model [@li2020behrt]. This includes converting raw EHR data into tokenised inputs by mapping vocabulary to numerical tokens and aligning patient histories into temporally ordered sequences of medical concepts. Each concept is paired with temporal features such as visit timestamps (positions), patient age, and visit-level segment encodings. Additional static features, including date of birth (DOB), gender, and optionally date of death (DOD), are prepended to each sequence. Separator and classification tokens can be optionally included. Numeric values can be binned and added as categorical tokens. Aggregation of similar concepts is supported via regex-based grouping, and the removal of specific concepts is also included via regex.
 
