### 2025-09-27 2c75867
Paper (#309)
* draft for joss paper

* figure

* minor changes

* error fix

* pdf generation

* figure sizes

* figure width

* improved image

* fixed reference

* name change

* transformer model

* minor comments

* format

* jpg -> png

* fixed caption

* white pipeline fig

* typo

* fix pipeline fig

* fix figure

* png from jpg
