### 2025-10-07 #8869 @espottesmith
Please check the missing DOIs.

Oh, also, there's a slight affiliation difference between archive and paper: "University of Copenhagen" vs. "Copenhagen University"
