+++ paper/paper.bib
@@ -2,13 +2,16 @@
   title={Core-behrt: A carefully optimized and rigorously evaluated behrt},
   author={Odgaard, Mikkel and Klein, Kiril Vadimovic and Thysen, Sanne M{\o}ller and Jimenez-Solem, Espen and Sillesen, Martin and Nielsen, Mads},
   journal={arXiv preprint arXiv:2404.15201},
-  year={2024}
+  year={2024},
+  doi={10.48550/arXiv.2404.15201}
 }
 
 @article{kolo2024meds,
   title={MEDS Decentralized, Extensible Validation (MEDS-DEV) Benchmark: Establishing Reproducibility and Comparability in ML for Health},
   author={Kolo, Aleksia and Pang, Chao and Choi, Edward and Steinberg, Ethan and Jeong, Hyewon and Gallifant, Jack and Fries, Jason A and Chiang, Jeffrey N and Oh, Jungwoo and Xu, Justin and others},
-  year={2024}
+  year={2024},
+  url={https://openreview.net/pdf?id=DExp3tRRel},
+  note = {Accessed: 2025-10-08}
 }
 
 @article{li2020behrt,
@@ -19,7 +22,8 @@
   number={1},
   pages={7155},
   year={2020},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+  doi={https://doi.org/10.1038/s41598-020-62922-y}
 }
 
 @article{warner2024smarter,
@@ -27,13 +31,15 @@
   author={Warner, Benjamin and Chaffin, Antoine and Clavi{\'e}, Benjamin and Weller, Orion and Hallstr{\"o}m, Oskar and Taghadouini, Said and Gallagher, Alexis and Biswas, Raja and Ladhak, Faisal and Aarsen, Tom and others},
   journal={arXiv preprint arXiv:2412.13663},
   year={2024}
+  doi{https://doi.org/10.48550/arXiv.2412.13663}
 }
 
 @article{touvron2023llama,
   title={Llama: Open and efficient foundation language models},
   author={Touvron, Hugo and Lavril, Thibaut and Izacard, Gautier and Martinet, Xavier and Lachaux, Marie-Anne and Lacroix, Timoth{\'e}e and Rozi{\`e}re, Baptiste and Goyal, Naman and Hambro, Eric and Azhar, Faisal and others},
   journal={arXiv preprint arXiv:2302.13971},
-  year={2023}
+  year={2023},
+  doi={https://doi.org/10.48550/arXiv.2302.13971}
 }
 
 @article{brown2020language,
@@ -42,7 +48,8 @@
   journal={Advances in neural information processing systems},
   volume={33},
   pages={1877--1901},
-  year={2020}
+  year={2020},
+  doi={https://doi.org/10.48550/arXiv.2005.14165}
 }
 
 @inproceedings{devlin2019bert,
@@ -50,14 +57,16 @@
   author={Devlin, Jacob and Chang, Ming-Wei and Lee, Kenton and Toutanova, Kristina},
   booktitle={Proceedings of the 2019 conference of the North American chapter of the association for computational linguistics: human language technologies, volume 1 (long and short papers)},
   pages={4171--4186},
-  year={2019}
+  year={2019},
+  doi={https://doi.org/10.48550/arXiv.1810.04805}
 }
 
 @article{pang2024cehr,
   title={CEHR-GPT: Generating electronic health records with chronological patient timelines},
[…]
