+++ paper/paper.bib
@@ -30,8 +30,8 @@
   title={Smarter, better, faster, longer: A modern bidirectional encoder for fast, memory efficient, and long context finetuning and inference},
   author={Warner, Benjamin and Chaffin, Antoine and Clavi{\'e}, Benjamin and Weller, Orion and Hallstr{\"o}m, Oskar and Taghadouini, Said and Gallagher, Alexis and Biswas, Raja and Ladhak, Faisal and Aarsen, Tom and others},
   journal={arXiv preprint arXiv:2412.13663},
-  year={2024}
-  doi{https://doi.org/10.48550/arXiv.2412.13663}
+  year={2024},
+  doi={https://doi.org/10.48550/arXiv.2412.13663}
 }
 
 @article{touvron2023llama,
