### 2025-10-13 1cb3163
Paper (#311)
* draft for joss paper

* figure

* minor changes

* error fix

* pdf generation

* figure sizes

* figure width

* improved image

* fixed reference

* name change

* transformer model

* minor comments

* format

* jpg -> png

* fixed caption

* white pipeline fig

* typo

* fix pipeline fig

* fix figure

* png from jpg

* small fixes

* typo

* updated pyproject

* removed prefix

* different layout
