+++ paper/paper.bib
@@ -23,7 +23,7 @@
   pages={7155},
   year={2020},
   publisher={Nature Publishing Group UK London},
-  doi={https://doi.org/10.1038/s41598-020-62922-y}
+  doi={10.1038/s41598-020-62922-y}
 }
 
 @article{warner2024smarter,
@@ -31,7 +31,7 @@
   author={Warner, Benjamin and Chaffin, Antoine and Clavi{\'e}, Benjamin and Weller, Orion and Hallstr{\"o}m, Oskar and Taghadouini, Said and Gallagher, Alexis and Biswas, Raja and Ladhak, Faisal and Aarsen, Tom and others},
   journal={arXiv preprint arXiv:2412.13663},
   year={2024},
-  doi={https://doi.org/10.48550/arXiv.2412.13663}
+  doi={10.48550/arXiv.2412.13663}
 }
 
 @article{touvron2023llama,
@@ -39,7 +39,7 @@
   author={Touvron, Hugo and Lavril, Thibaut and Izacard, Gautier and Martinet, Xavier and Lachaux, Marie-Anne and Lacroix, Timoth{\'e}e and Rozi{\`e}re, Baptiste and Goyal, Naman and Hambro, Eric and Azhar, Faisal and others},
   journal={arXiv preprint arXiv:2302.13971},
   year={2023},
-  doi={https://doi.org/10.48550/arXiv.2302.13971}
+  doi={10.48550/arXiv.2302.13971}
 }
 
 @article{brown2020language,
@@ -49,7 +49,7 @@
   volume={33},
   pages={1877--1901},
   year={2020},
-  doi={https://doi.org/10.48550/arXiv.2005.14165}
+  doi={10.48550/arXiv.2005.14165}
 }
 
 @inproceedings{devlin2019bert,
@@ -58,7 +58,7 @@
   booktitle={Proceedings of the 2019 conference of the North American chapter of the association for computational linguistics: human language technologies, volume 1 (long and short papers)},
   pages={4171--4186},
   year={2019},
-  doi={https://doi.org/10.48550/arXiv.1810.04805}
+  doi={10.48550/arXiv.1810.04805}
 }
 
 @article{pang2024cehr,
@@ -66,7 +66,7 @@
   author={Pang, Chao and Jiang, Xinzhuo and Pavinkurve, Nishanth Parameshwar and Kalluri, Krishna S and Minto, Elise L and Patterson, Jason and Zhang, Linying and Hripcsak, George and G{\"u}rsoy, Gamze and Elhadad, No{\'e}mie and others},
   journal={arXiv preprint arXiv:2402.04400},
   year={2024},
-  doi={https://doi.org/10.48550/arXiv.2402.04400}
+  doi={10.48550/arXiv.2402.04400}
 }
 
 @inproceedings{pang2021cehr,
@@ -76,7 +76,7 @@
   pages={239--260},
   year={2021},
   organization={PMLR},
-  doi={https://doi.org/10.48550/arXiv.2111.08585}
+  doi={10.48550/arXiv.2111.08585}
 }
 
 @article{gu2023mamba,
@@ -84,7 +84,7 @@
   author={Gu, Albert and Dao, Tri},
   journal={arXiv preprint arXiv:2312.00752},
   year={2023},
-  doi={https://doi.org/10.48550/arXiv.2312.00752}
+  doi={10.48550/arXiv.2312.00752}
 }
 
 @article{rasmy2021med,
@@ -96,7 +96,7 @@
   pages={86},
   year={2021},
   publisher={Nature Publishing Group UK London},
-  doi={https://doi.org/10.1038/s41746-021-00455-y}
+  doi={10.1038/s41746-021-00455-y}
 }
 
[…]
