### 2025-10-19 468f9a5
updated references
