+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{odgaard2024core,
-  title={Core-behrt: A carefully optimized and rigorously evaluated behrt},
+  title={Core-BEHRT: A carefully optimized and rigorously evaluated BEHRT},
   author={Odgaard, Mikkel and Klein, Kiril Vadimovic and Thysen, Sanne M{\o}ller and Jimenez-Solem, Espen and Sillesen, Martin and Nielsen, Mads},
   journal={arXiv preprint arXiv:2404.15201},
   year={2024},
@@ -110,14 +110,18 @@
   doi={10.48550/arXiv.1912.01703}
 }
 
-@misc{huggingface_transformers,
-  title = {Transformers},
-  howpublished = {\url{https://huggingface.co/docs/transformers/en/index}},
-  note = {Accessed: 2025-09-25}
+@online{huggingface_transformers,
+  title   = {Transformers},
+  author  = {{Hugging Face}},
+  year    = {2025},
+  url     = {https://huggingface.co/docs/transformers/en/index},
+  urldate = {2025-09-25}
 }
 
-@misc{mlflow,
-  title = {mlflow.azureml},
-  howpublished = {\url{https://mlflow.org/docs/1.22.0/python_api/mlflow.azureml.html}},
-  note = {Accessed: 2025-09-25}
+@online{mlflow,
+  title   = {mlflow.azureml},
+  author  = {{MLflow}},
+  year    = {2025},
+  url     = {https://mlflow.org/docs/1.22.0/python_api/mlflow.azureml.html},
+  urldate = {2025-09-25}
 }
